requirements.txt
fastapi
uvicorn
pymongo
pydantic
certifi
cryptography
spacy
python-multipart
pdfminer.six
pandas
python-docx
openpyxl

